split commitment status calculator unit tests into transfer and non-transfer testfixtures
